refactor(storage): own record side-effect persistence
